+++ cotk/dataloader/sentence_classification.py
@@ -71,7 +71,7 @@ class SentenceClassification(LanguageProcessingBase):
 		res = {}
 		batch_size = len(indexes)
 		res["sent_length"] = np.array( \
-			list(map(lambda i: len(self.data[key]['sent'][i]), indexes)))
+			list(map(lambda i: len(self.data[key]['sent'][i]), indexes)), dtype=int)
 		res_sent = res["sent"] = np.zeros( \
 			(batch_size, np.max(res["sent_length"])), dtype=int)
 		res["label"] = np.zeros(batch_size, dtype=int)
